Move logic to retrieve instance properties to common.utils
This logic will be used by sqswatcher as well
